+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 - **API**
   - Add `/sdapi/v1/lora?lora=<lora_name>` endpoint that returns full lora info and metadata  
+  - Add `/sdapi/v1/controlnets?model_type=<model_type|all|None>` endpoints that returns list of available controlnets for specific model type  
 
 - **Fixes**  
   - IPEX with DPM2++ FlowMatch samplers